Old config reinstated for now.
Pull requests as was would mean that we didn't read from the old queues, so whilst the new code writes to the new "notify" queue, the old code would write to the old plethera of queues. And the readers may not pick it up depending on the order of the deploy.

Safer to leave readers reading from all queues until after the deploy then doing a second deploy to tidy up the config.