add new notification_status_types static table
enums are hard to develop with - it is a slow and unchunkable change
to add a new type or remove an old type. We're slowly moving away from
using enums, in favour of static type tables, where the name is the
primary key.

This commit contains an alembic script that creates that table, and
creates a new nullable column on notifications and
notification_history that refer to it.