Make missing personalisation error consistent
In other places the text we use for this error message is "Missing personalisation: name, date, thing". See:
- https://github.com/alphagov/notifications-api/blob/72b108b694c76a4559be04cb51c468be8db46c57/app/template/rest.py#L125
- https://github.com/alphagov/notifications-api/blob/717c0510a3986123497d0e49dd52ae3f87d953d3/app/notifications/rest.py#L206
- https://github.com/alphagov/notifications-api/blob/05a179c6efb051de1ea70bd560d7cb6966c44ff6/app/v2/template/post_template.py#L38

For some reason this part of the codebase says "Template missing personalisation: …". This is inconsistent, and also confusing because it’s the API call that’s missing the personalisation, not the template itself. 

This commit changes the error message to be consistent with the majority of the codebase, which uses the less confusing wording.